+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to dbapi will be documented in this file
 
+Date: 2015-4-19
+
+### Added
+- Table Deletion
+
 Date: 2015-4-16
 
 ### Added